person controller, unit test for convention on disabled behavior when false
