+++ PersonsAPI/Models/Repositories/Interfaces/IPersonRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace PersonsAPI.Models.Repositories.Interfaces
+{
+   public interface IPersonRepository
+    {
+        
+    }
+}
