+++ PersonsAPI/Models/Repositories/MemoryPersonRepository.cs
@@ -1,0 +1,13 @@
+﻿using PersonsAPI.Models.Repositories.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace PersonsAPI.Models.Repositories
+{
+    public class MemoryPersonRepository: IPersonRepository
+    {
+
+    }
+}
